+++ stable_baselines/ddpg/ddpg.py
@@ -1039,7 +1039,7 @@ class DDPG(OffPolicyRLModel):
 
         return actions, None
 
-    def action_probability(self, observation, state=None, mask=None, actions=None):
+    def action_probability(self, observation, state=None, mask=None, actions=None, logp=False):
         observation = np.array(observation)
 
         if actions is not None: