+++ stable_baselines/her/her.py
@@ -126,8 +126,8 @@ class HER(BaseRLModel):
     def predict(self, observation, state=None, mask=None, deterministic=True):
         return self.model.predict(self._check_obs(observation), state, mask, deterministic)
 
-    def action_probability(self, observation, state=None, mask=None, actions=None):
-        return self.model.action_probability(self._check_obs(observation), state, mask, actions)
+    def action_probability(self, observation, state=None, mask=None, actions=None, logp=False):
+        return self.model.action_probability(self._check_obs(observation), state, mask, actions, logp)
 
     def _save_to_file(self, save_path, data=None, params=None):
         # HACK to save the replay wrapper